Extracted out download function
